fix(assert): Move Actions into assert mod
Cherry pick 125d7c4ba3754dab55f1ecfc030184b56868cc710 (#294)
